feat(RadicalModes):Create a configurable mode
